refactor: Update class names for grpc communication
