+++ p2pfl/communication/grpc/grpc_client.py
@@ -20,7 +20,7 @@ from typing import List, Optional, Union
 
 import grpc
 
-from .neighbors import GrpcNeighbors
+from p2pfl.communication.grpc.grpc_neighbors import GrpcNeighbors
 from p2pfl.communication.grpc.proto import node_pb2, node_pb2_grpc
 from p2pfl.management.logger import logger
 from p2pfl.settings import Settings
