+++ p2pfl/communication/grpc/grpc_neighbors.py
@@ -1,3 +1,20 @@
+#
+# This file is part of the federated_learning_p2p (p2pfl) distribution (see https://github.com/pguijas/federated_learning_p2p).
+# Copyright (c) 2022 Pedro Guijas Bravo.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, version 3.
+#
+# This program is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+
 import time
 from typing import Optional, Tuple
 
@@ -52,7 +69,7 @@ class GrpcNeighbors(Neighbors):
 
             # Add neighbor
             return (channel, stub, time.time())
-        
+
         except Exception as e:
             logger.info(self.self_addr, f"Crash while adding a neighbor: {e}")
             # Re-raise exception
