+++ p2pfl/examples/mnist.py
@@ -119,4 +119,4 @@ if __name__ == "__main__":
     # Settings
     set_test_settings()
     # Launch experiment
-    test_convergence(25, 25, epochs=0)
+    test_convergence(2, 2, epochs=0)
